Reformat codes
