+++ lib/pages/home_page.dart
@@ -1,0 +1,24 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          const Text('Login successed'),
+          ElevatedButton(onPressed: _logout, child: const Text('Logout')),
+        ],
+      ),
+    );
+  }
+
+  void _logout() {
+    FirebaseAuth.instance.signOut();
+  }
+}
